+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @swapkit/tokens
 
+## 4.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`c048f39`](https://github.com/swapkit/SwapKit/commit/c048f39a74b21182d2093e97883c1dc8971293f3)]:
+  - @swapkit/helpers@4.0.13
+  - @swapkit/wallets@4.1.12
+  - @swapkit/toolboxes@4.0.13
+
+## 4.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/wallets@4.1.11
+  - @swapkit/helpers@4.0.12
+  - @swapkit/toolboxes@4.0.12
+
 ## 4.0.11
 
 ### Patch Changes