+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swapkit/ui
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/browser@4.0.14
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/browser@4.0.13
+
 ## 0.1.12
 
 ### Patch Changes