+++ packages/wallet-hardware/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @swapkit/wallet-hardware
 
+## 4.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`c048f39`](https://github.com/swapkit/SwapKit/commit/c048f39a74b21182d2093e97883c1dc8971293f3)]:
+  - @swapkit/helpers@4.0.13
+  - @swapkit/toolboxes@4.0.13
+
+## 4.1.11
+
+### Patch Changes
+
+- [#1478](https://github.com/swapkit/SwapKit/pull/1478) [`85e0cea`](https://github.com/swapkit/SwapKit/commit/85e0ceafdec00c4be7dd6cfefacce360437a4e51) Thanks [@towanTG](https://github.com/towanTG)! - Fixes Ledger Zcash implementation
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.12
+  - @swapkit/toolboxes@4.0.12
+
 ## 4.1.10
 
 ### Patch Changes